Refactoring of label_t into input_label_t, channel_label_t and src_label_t
The mixer input can be either a logical input or an output channel.
Therefore originally we had one label_t list that described
all possible mixer inputs. We then used subtraction to determine
whether a mixer input refers to a logical input or an output channel.
This lead to problems in the UI as now the UI would need information
about what label_t values are allowed to show. E.g. a mixer destination
can only be an output channel, while a mixer unit Switch can only
be a logical input.

The refactoring now defines separate lists for logical inputs and
for the output channels. We then created a src_label_t list
that holds the combined values of both input_label_t and
channel_label_t. For each of the sub-lists we define their
first and last value as a #define so that we can range check
to determine which sub-list a mixer source belongs, and redirect
retrieving values to the appropriate module.