When downloading a new model set the RF address to a wrong value
While downloading a new model, the current model may still be
connected. The new model data may not make sense for the current
model, but the transmitter already executes the mixer and other
settings, which may cause the model to run off or do other
weird things.

So before we download we set the RF address to a random value,
which causes the transmitter to send data to the current model.
We then download the new model. The correct RF address is the last
part of the model data, so once all model data is set the transmitter
will receive the correct address and everything goes smooth