Limit to two packets in flight over the websocket
If we are sending packets too quickly, the browser combines multiple
WebSocket packets in one TCP packet. It seems that the ESP Async
Websocket library can not handle that and silently ignores
the packet, resulting in lost data.

The work-around is to send packets slowly for now. Not ideal,
but at least works.

We also remember the last connected bridge URL, so that
we try that first when we connect again during the same session.
Furthermore, we give the ESP configurator higher priority than
the hostname of the page since this will be most common use-case
in actual use.